document-analysis-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
kafka-python==2.0.2
minio==7.2.0
mistralai==1.9.11
python-dotenv==1.0.0
pydantic>=2.10.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
python-docx==1.1.0
PyPDF2==3.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
